// src/app/shared/models/app-routes.ts
export enum AppRoutes {
    ROOT = '/',
    REGISTER = 'register',
    LOGIN = 'login',
    CHANGE_PASSWORD_AFTER_LOGIN="login/changepassword",
    FORGOT_PASS = 'forgotpassword',
    NOTIFICATION = 'notification',
    REGISTER_PHARMACY = 'registerpharmacy',
    OTP_VERIFY = 'otpverification',
    RESET_PASS = 'resetpassword',
    DASHBOARD ='dashboard',
    NEW_REQ='newrequest',
    DOSE_ONE='doseone',
    DOSE_TWO='dosetwo',
    USER_VAC='uservacform',
    ADD_OTH_DOSE='addotherdose',
    DOSE_HISTORY='dosehistory',
    FACILITY_USER_VAC = 'facilityuservacform',
    FACILITY_ADD_OTH_DOSE = 'facilityaddotherdose',
    FACILITY_DOSE_HISTORY = 'facilitydosehistory',
    VACCINATOR = 'vaccinator',
    ADD_VACCINATOR = 'addvaccinator',
    SETTINGS='settings',
    REMINDER='reminder',
    PROFILE = 'profile',
    USERS = 'users',
    FACILITIES = 'facilities'
}